\name{batter_transition}
\alias{batter_transition}

\title{
  Count Transitions for a Batter
}
\description{
  Computes and graphs count transitions for a specific batter
}
\usage{
  batter_transition(d, name)
}
\arguments{
  \item{d}{
    Retrosheet play-by-play data frame
  }
  \item{name}{
    Full name of a batter
  }
}
\value{
  List with components p_trans, p_prob, and plot
}
\author{
  Jim Albert
}


