\name{compute.runs.expectancy}
\alias{compute.runs.expectancy}

\title{
Computes Run Expectancies from Retrosheet Data
}
\description{
Adds header and computes run expectancies from a Retrosheet play-by-play data frame in the current working directory
}
\usage{
  compute.runs.expectancy(season)
}
\arguments{
  \item{season}{
    Season of interest
  }
}
\value{
  Data frame of Retrosheet data for season with additional variables added
}
\author{
  Jim Albert
}


