\name{PA_count}
\alias{PA_count}

\title{
  Length of Plate Appearance
}
\description{
  Defines new variable giving sequence of balls and strikes
}
\usage{
  PA_count(d)
}
\arguments{
  \item{d}{
    Retrosheet play-by-play data frame
  }
}
\value{
  Retrosheet data frame with new variables pseq, Count, and pseq_length
}
\author{
  Jim Albert
}


