\name{parse.retrosheet2.pbp}
\alias{parse.retrosheet2.pbp}

\title{
Download Retrosheet play-by-play files
}
\description{
Downloads and parses Retrosheet play-by-play files for a given season
}
\usage{
  parse.retrosheet2.pbp(season)
}
\arguments{
  \item{season}{
    Season of interest
  }
}
\value{
  Data frame of Retrosheet data for season
}
\author{
  Jim Albert
}


