\name{pitch_count_graph}
\alias{pitch_count_graph}

\title{
  Pitch Count Graph
}
\description{
  Constructs graph showing the mean runs value of plate appearances going through each possible count
}
\usage{
  pitch_count_graph(d_season)
}
\arguments{
  \item{d}{
    Retrosheet play-by-play data frame with variables c01, c10, ...
  }
}
\value{
  graph showing the mean runs value of plate appearances going through each count
}
\author{
  Jim Albert
}


