\name{pitcher_transition}
\alias{pitcher_transition}

\title{
  Count Transitions for a Pitcher
}
\description{
  Computes and graphs count transitions for a specific pitcher
}
\usage{
  pitcher_transition(d, name)
}
\arguments{
  \item{d}{
    Retrosheet play-by-play data frame
  }
  \item{name}{
    Full name of a pitcher
  }
}
\value{
  List with components p_trans, p_prob, and plot
}
\author{
  Jim Albert
}


