\name{prob_transitions}
\alias{prob_transitions}

\title{
  Probability Count Transitions
}
\description{
  Constructs the probabilities of all count transition in a plate appearance
}
\usage{
  prob_transitions(dt)
}
\arguments{
  \item{dt}{
    Data frame giving beginning and ending counts for each pitch
  }
}
\value{
  data frame of the probability of all count transitions
}
\author{
  Jim Albert
}


