\name{setup_thru_counts}
\alias{setup_thru_counts}

\title{
  Sets Up Pass Through Count Variables
}
\description{
  Computes all of the pass through count variables in a plate appearance
}
\usage{
  setup_thru_counts(d)
}
\arguments{
  \item{d}{
    Retrosheet play-by-play data frame
  }
}
\value{
  data frame with new variables pseq, c01, c10, ...
}
\author{
  Jim Albert
}


