\name{strikes}
\alias{strikes}

\title{
  Number of Strikes in Count
}
\description{
  Finds the number of strikes in a count variable
}
\usage{
  strikes(count)
}
\arguments{
  \item{count}{
    Vector of counts in "b-s" format
  }
}
\value{
  vector of strikes in each count
}
\author{
  Jim Albert
}


