\name{transition_graph}
\alias{transition_graph}

\title{
  Graph of Count Transitions
}
\description{
  Constructs a graph of the count probability transition matrix
}
\usage{
  transition_graph(S, title = "")
}
\arguments{
  \item{S}{
    Data frame giving the count transition probabilities
  }
  \item{title}{
    Prefix string to add to the graph title
  }
}
\value{
  graph of the count transition probabilities
}
\author{
  Jim Albert
}


