\name{transition_work}
\alias{transition_work}

\title{
  Count Transitions
}
\description{
  Computes and graphs count transitions for a Retrosheet dataset
}
\usage{
  transition_work(d, title = "")
}
\arguments{
  \item{d}{
    Retrosheet play-by-play data frame
  }
  \item{title}{
    Title to add to graph
  }
}
\value{
  List with components p_trans, p_prob, and plot
}
\author{
  Jim Albert
}


